rewrite the scritp to mimic merging closely
Since there are no merge conflict at the moment, the script has been
rewritten to better mimic the merging process. It assumes the PR's
are merged in order, which of course isn't going to happen. At the
same time it simulates a rebase before merging, which creates a more
linear history, closer to what we're expecting to see in the future
(albeit with a different order).